input: rework decoder_device handling
Merge input_resource_GetVoutDecoderDevice() and input_resource_StartVout() into
input_resource_RequestVout().

RequestVout() can be used to create a vout without starting it (if cfg->fmt is
NULL). It allows the decoder to create a decoder_device from the vout manually,
then call RequestVout() again with a valid cfg->fmt to start it.