apple: build.sh: add an option to build dynamic objects
Add the --enable-shared flag in order to build each plugin and the
libvlc/libvlccore libraries as a dynamic object.

The goal is to be able to ship them in a bundle archived as an IPA,
which can then be signed and installed on real device without needing
complex additional linking steps and generating new files.