opengl: renderer: improve cube generation
The cube coordinates were defined one by one, giving little possibility
for readable refactoring or little explanation for those not knowing how
the cube is built.

Instead, provide a helper macro generating the coordinates and other
helper macros in order to select the correct face to generate. It will
ease future refactoring, in particular to rotate the cube.