opengl: support "blend" filters
A blend filter draws over the result of any previous filter.

Expose a flag to be set by filter modules implementation, and draw them
using the same draw framebuffer as the last (non-blend) filter.

If the first filter is a blend filter, then a "draw" filter (a non-blend
filter which just draws the input picture) is automatically inserted.

Co-authored-by: Alexandre Janniaux <ajanni@videolabs.io>
Co-authored-by: Maxime Meissonnier <mmeisson@outlook.fr>
Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>