opengl: lazy-load sampler
The sampler was created and passed to the filter in the Open() function.

But only non-blend filters need a sampler to read the input picture
(blend filters do not read it).

To avoid creating an unnecessary sampler (including GLSL code
generation), retrieve the sampler from an owner function, which creates
it on-demand.

Co-authored-by: Alexandre Janniaux <ajanni@videolabs.io>
Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>