avcodec: va: wait indefinitely until a surface is available
If there's no surface available the decoder should wait for one to be available,
not get an error and not try to decode anymore. When pictures (and thus
surfaces) were coming from the display a picture_pool_Wait() was used, so in
the normal case we should do something similar.

The infinite looping shouldn't be an issue as on Flush or Close the decoder
will release all the past pictures/surfaces. So there will be new surfaces
available for past calls.

We could improve the code by using a conditional variable or semaphore to
unlock the pending requests rather than using vlc_tick_sleep().