opengl: renderer: respect spherical-v2 for cubemap
The metadata signalling the usage of a cubemap are defined in the
spatial-media repository[1] and enforce a specific layout for layout
identifier 0, which is the only one specified.

Implement this layout instead of the previous one, which was Facebook
one (using Mongolfiere 360 sample), especially given that Facebook has
mostly gone to equirectangular 360 videos since then.

It fixes the definition of each face so that they match their natual
axis with the following coordinate system (x being forward, and z being
a vector to the left, the whole in right-hand convention):

       y     x
          \ |
           \|
   z <-----(+)

It now uses the following picture layout:

1
 +-------+-------+------+
 | right | left  | up   |
 +-------+-------+------+      v
 | down  | front | back |      |
 +-------+-------+------+      +-- u
0                        1

For testing purpose, one can generate a sample for this format by using
ffmpeg, for example by converting an equirectangular video into cubemap
using the v360 filter.

   -lavfi v360=input=equirect:output=c3x2

Matroska projection can also be set, for example with:

   mkvpropedit sample --edit track:v1 --set projection-type=2

[1]: https://github.com/google/spatial-media/blob/master/docs/spherical-video-v2-rfc.md#cubemap-projection-box-cbmp