filter: move pf_get_attachments to owner
Attachments are not coming from the filter but from the pipeline and is
called by the filter implementation itself. Thus, this is defined by the
owner implementation and should not be in the operation callbacks.

It's particularily noticeable in the callsite given that the
get_attachments implementation is using owner.sys.