configure.ac: use absolute path for protoc
When compiling for iOS, protoc is usually found in PATH but PATH is
defined by the build script, which mean only `protoc` will be stored in
the PROTOC variable. When recompiling using `make`, the PATH is not
defined anymore and it will fail running the command.

There's two solutions for this issue:
 - add extra/tools in the look-up path when calling AC_CHECK_PROGS
   instead of exporting it in PATH
 - use absolute path

The first option doesn't work because extra/tools is always in path in
the configure script or in the compile script, but is not when using
make directly.

The second option will work for this, and additionally will prevent
using the extra/tools one in a project that has been configured with the
system one for example.