lib: media: rewrite recursion into list exhaustion
As the comment described it, recursion could lead to a stack overflow
when parsing very long recursive playlist. Instead, use temporary
dynamic allocation to chain the addition requests and process them in
the same order.

The out-of-memory errors are only returned to the user as a log and the
preparsed playlist item is truncated, which is non-optimal.