aout: use a new chan_mode for binaural
Instead of using a custom cfg with an extra "audio filter".

i_chan_mode is already used by "audio converter" for dolbystereo.

The spatialaudio "audio filter" is converted to an "audio converter".
Such type of modules are loaded automatically when a conversion is needed
(the chan_mode conversion). Since this module is now an "audio
converter", don't override in/out fmt, but test for in/out fmt
compatibility.