vout: kms: don't use a custom picture destroy
The picture is not used outside of this file. We don't need to use the picture
release refcount to release the resources used in the picture.

That's how other display modules handle their dummy local picture. The picture
is allocated locally and the resources are released locally at the end.