auhal: fix wrong host time with SPDIF render callbacks
This caused a freeze as the aout returned infinite values from time_get() with
SPDIF output.

Indeed, the inNow time is in the past (time when the callback is executed) and
a substraction with mach_absolute_time() returned a negative value.

The issue is revealed by 303d12153f67742ac21a31efb1eeb1f4aaa04f08 but the
SPDIF host time had always been wrong.