macosx: update EQ frequency labels according to VLC frequency setting
An option exists for selecting between VLC frequency bands and ISO frequency
bands for the audio equalizer; however, as of this writing, the text labels in
the macOS interface do not quite reflect the selected option.  (For example, if
the option to use ISO frequencies is selected in preferences and VLC is
restarted, VLC frequencies still appear in the effects interface.)  This patch
introduces a change that dynamically updates the frequency labels according to
the VLC frequency setting.

Signed-off-by: Marvin Scholz <epirat07@gmail.com>