decoder: fix pause+close deadlock for decoders waiting internally for buffers
When paused the decoder decodes until it exhausts its available buffers then
waits until new buffers are available. When seeking we force flushing the vout
which releases the buffers the decoder is waiting for.

When closing while pause we didn't flush the vout so the decoder was still
waiting for buffers even though it's not going to use them (or display them).
So we do the same flushing to automatically release the used/pending buffers.