vout: direct3d11: clean the wait until all the commands are processed
We need to unlock the device context while we wait until it's finished with our
commands. As it may block all decoder threads for 2ms or (a lot) more. The
decoder may have a chance to send some decoding commands and we may wait longer
when locking back the device context. This won't have much influence on when
the picture is finished rendering as it won't wait for potential new commands
that have been sent to the video context.

We don't need to lock the device context in PreparePicture, it's already locked.

Use a ID3D11Asynchronous pointer directly. We don't need to know the underlying
type, except when creating it.