macosx: add VLCSystemVolume class
This is a convenience class, based on the NSSound+VLCAdditions
extension, with the goal to replace it.

There is not much benefit to keep it an extension of NSSound as there
is nothing that this class needs from NSSound to function at all, so
not cluttering the NSSound namespace seems like a better approach here.