direct3d11: wait indefinitely for the rendering to be done
If we don't, D3D11 complains that we didn't wait for the finished result
matching a ID3D11DeviceContext::End() call.
As long as GetData() returns S_FALSE we can safely wait. Any other value will
exit the loop.

We want the rendering time seen by the vout to be as accurate as possible
rather than stop waiting after some time and pretend the rendering happened
when it didn't.
The granularity of the wait should be as small as possible to have the
rendering time stats as accurate as possible.