video_output: use filter lock to change filters in the vout thread
We don't need to wait when enabling filters, it's OK if we miss the current
frame and apply it in the next loop iteration. The filter changes will
be applied just before the picture is rendered, only once between frames. It is
now done after the control pop deadline, so will eat some of the rendering
time.

We don't use the control lock anymore (from vout_control_PushString), just the
filters lock.

Now the filter changes are applied even in frame by frame mode.