lua/http: ask user before sending web interface URL to Google QR code API
The existence and URL of the web interface of a VLC instance can be
sensitive information, and transmitting it over the internet to a third
party might not be something that the user is okay with.

This creates a new dialog informing the user and prompting them for
confirmation, before sending the URL to a remote web API to generate a
QR code for it.

Refs #25005