common.lua: fix utility resolving dot segments in paths
This utility didn't work at all on Windows because it had no support for
'\', it also had glaring bugs and shortcomings, and its implementation
was broken. Lua has no knowledge of the OS and directory separator
currently in use, so this takes the one sensible approach here:
supporting URLs to let the utility operate on unambiguous URL paths
rather than file paths.