spu: rss: fix first character of display being omitted
The variable p_sys->i_cur_char is used to track how many chars shall be left
out when printing a title into the display buffer. If this variable is
initialized to zero, then in the very first iteration it is already set to 1 in
function *Filter. This then leads to the effect that the very first char is
never printed.