spu: rss: initialize title strings with non NULL pointer
If no valid title could be parsed out of an RSS feed, the segmentation fault
caused VLC to crash. The reason therefore was that the expected pointers to the
char buffers containing the titles were NULL pointers.

This commit initializes the titles of both the feeds and all items with a zero
string.

Additionally, the check whether already a title was found during parsing is
removed as it is not reasonable why a RSS stream should provide several title
tags for one item.