macosx/sprefs: Constraints enhancements for MediaLibrary View
Right now, the MediaLibraryView is having wrong placements of button layer viz.
Add Folder, Remove Folder, Ban Folder, this commit fixes that. So the buttons
are showing within the listview of the items, which is not the intented way.

Also fixes the Video View's misplaced views  and overfitting of buttons.

Signed-off-by: Felix Paul Kühne <felix@feepk.net>