opengl: yadif: compute lines in an intermediate buffer
Add an intermediate framebuffer of size height / 2 so as to store the
computation of the lines with the yadif algorithm. Since it's expensive
and only needed for one line over two, it improves the fragment shader
performances by trading bandwidth instead, and also has better latency
hiding in the gathering step.

It's particularily helpful on tiled based deferred architectures.