spu: rss: Fix TODOs: If no item exists then RSS feed is skipped
In function *Filter, there already existed a TODO note that it needs to be
checked whether the feeds have items at all. As this can change with every
parsing of the URLs, this can only be done dynamically at that point in time
when the display is rendered.

The TODO note in the function comment of ParseUrls seems to be outdated as the
check for an non empty url string is already done in function CreateFilter() in
lines 263..268.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>