lua/http: avoid diverging legacy code to print status.xml
The constraint with ES information was that it produced invalid XML,
but now that this is fixed, there is no reason to exclude standard data
from the output. In particular this adds support for chapter and title
information, and any future new field, while actually reducing the
amount of specific code.

Fixes #10443