kms: use the source format to create the local picture
At this stage the only difference betweeen vd->source and vd->fmt is the SAR.
Since this local picture is only use to copy planes, the SAR is never used.

It's possible that the original code is wrong as it doesn't take in account the
"kms-vlc-chroma" forced chroma that the core will assume is used.