display: don't force a 0 SAR anymore on the display preferred format
In some cases the display modules may rotate the video format, resulting in
inverted and non matching SAR. But in that case the width/height will also be
rotated and thus not matching either.

Also the comparision between the input and preferred format of the display
module explicitly zero the SAR for comparison (although it may hide some needed
conversion).