qt: fix incorrect param type
which was causing a whole bunch of "'< 0' is always false" warnings.

review of the original fix ([1]) indicated that in fact the parameter was
wrong rather than the checks within the function being out of step with
the final design of the function interface, since apparently callers
are in fact passing in a signed index, where a negative represents an
invalid index which the function is expected to handle appropriately.

[1]: https://mailman.videolan.org/pipermail/vlc-devel/2020-September/137711.html

Co-authored-by: Pierre Lamot <pierre@videolabs.io>
Signed-off-by: Pierre Lamot <pierre@videolabs.io>