vlc_list: prevent name conflict
The macro vlc_list_foreach() declares an internal variable
vlc_list_it_XXX variable for looping, where XXX is the variable provided
by the user to receive the item.

If this name is "next", the resulting internal variable name is
"vlc_list_it_next", which conflicts with the function
vlc_list_it_next().

Add an _ to the internal variable name to prevent it.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>