picture_chain: modify the append to take care of the chain init
By default p_next is always set to NULL in a picture. We can always use it as
a proper picture chain.

When a chain becomes empty p_next goes back to zero and the "known tail" is
invalid. But we only use the tail when the front is not empty, so there is no
problem.