vlc_spawn: vector posix_spawn/fork code
This adds vlc_spawn(), vlc_spawnp() and vlc_waitpid() to factor the
common code around the posix_spawn() family. It is added as part of
the plugin API so that core code can use it (namely vlc_getProxyUrl()),
which would not be possible with a shared C source in modules/.

Win32 back-end may be provided later, e.g. based on old vlc_execve()
implementation, if/when this actually gets used on Windows.