config: display unknown options
When an unknown option is requested, the assertion triggers on the null
configuration variable, but the name of the missing variable is not
displayed.

While it's (sometimes) possible to display it through debugger or
additional code, the assertion is very usual and the message is
especially useful when debugging mobile devices or using libvlc without
being a core developer.