transcode: remove the converter between the decoder/prerender and user filters
Just like in the video output, the user filters should adapt to the source,
better than they should adapt to the encoder. In particular, hardware support
will be favored in more cases.

Just like in the video output, the last converter chain is responsible to adapt
the decoded + prerendered + filtered pictures to the encoder requested input
format (see osys->converters display.c).

In the end we have the same filters as the video output: static, user, final
converter.

With the orientation changes done outside of the p_conv_nonstatic, this filter
cahin was always NULL (b_orient always false).