demux: always use strongly-typed file extensions
There are two sets of file extension to demux module mapping:
- The "strong" set contains file extensions deemed reliable, either
  because the demux probe is reasonably strict and safe, or because
  the extension is rarely mismatched.
- The "quick" set contains file extensions that are not considered
  reliable. This is essentially "mp3" which is supposedly (?) often
  misused for non-MPGA (or non-raw ES) files.

The quick set cannot be used in normal parsing, as it could pick the
the wrong demuxer (that is to say, MPEG ES). It was introduced in the
original preparsing changeset, presumably to limit the overhead.

There are however no obvious reasons not to use the strong set for
preparsing. This patch does exactly that, and then removes duplicated
mappings for Ogg and WMA.