modules: activation function cannot be NULL
There would be no ways to use the module. It couldn't even provide
further callbacks. There are indeed no modules with NULL callbacks
at the moment.

As an exception, the LibVLC core "plugin" has a dummy module without
activation callback, but also without capability, so it cannot be
activated anyway. (This is really a design bug in the plugin descriptor
syntax which cannot represent a plugin containing zero modules.)