ytdl: extract playlist non-recursively ("flat")
This uses the extract-flat mode of YoutubeDL, which skips parsing
individual items within a playlist. We define a dedicated MRL scheme
to track a YoutubeDL playlist item, and parse it only when the item is
actually opened.

This has two benefits:
1) Extracting a playlist is dramatically faster.
2) Expiring media URL can be played even if the playlist is long.

This does *not* solve the remaining problem that expiring URLs cannot
be saved and replayed later.