opengl: fix framebuffer initialization
The OpenGL offscreen implementation binds its draw framebuffer on open,
and expects the caller to draw on this current framebuffer, then calls
vlc_gl_Swap().

However, on initialization, the OpenGL video filter (as filter_t) first
loads the offscreen implementation, then initializes the OpenGL filters
(vlc_gl_filters), which may change the current bindings. As a
consequence, a wrong framebuffer may be bound on the first call to
Filter(), so the first picture is broken.

To avoid the problem, restore the bindings on filters initialization.