opengl: yadif: fix first pictures
The commit "opengl: yadif: render first frames correctly" was wrong.

Firstly, probably due to multiple refactors, the local variables were
never used after they were updated (so in fact the code did nothing).

Secondly, to rotate the buffers, sys->next must be assigned to
(sys->next + 1) % 3. Since this was always the value of prev, sys->next
was directly assigned from prev.

This assumption is broken once the first problem is fixed, because
prev is updated for the two first frames to be same same as cur or next.
As a consequence, to avoid breaking the framebuffer rotation, next must
not be computed from the prev value.