po: purge crystal hd artefacts
commit e1a90aa943b55734cdf8f25f5e2dc12a547b0c30 removed an
entry from po/vlc.pot which consequently causes builds to
update all of the po files accordingly (in many cases moving
existing translations to commented out copies further down
in the files), which then sit as changed files in the
unstaged list.

this purges the left over artefactsto stop this being a problem.

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>