filter: declare the filter function called in video filter wrappers
By default it declares a static function. There are new variants for the rare
cases where the function is not static.

For now the function to call cannot return an error to match the existing usage.

Also declare the close function to make sure there's more warnings when it's
used with the wrong callback signature.