coreaudio: fix play of uninitialized data (loud CRACK)
When starting deferred (likely), ca_Render() is filling the output
buffer with 0s (silence) until the requested start time is reached. When
the host time is near the requested start time, the output buffer is
partially filled with 0s, and partially filled with valid data.

In that particular case, the output buffer offset was not updated
causing the valid data to be copied at the beginning of the output
buffer, leaving some uninitialized data at the end of the buffer.

Fixes #25142

Signed-off-by: Marvin Scholz <epirat07@gmail.com>
(cherry picked from commit 61037b63695182fc2031614ef554d000ce5c9a0d)
Signed-off-by: David Fuhrmann <dfuhrmann@videolan.org>