videotooolbox: add fallback option
Add an option to control the error handling behaviour of Videotoolbox.
Usually, it was triggering a fallback (typically on avcodec) every time
an error happened within the played stream, which can happen when
corrupting the input stream. It was done to protect the decoder against
non-playable files for which the decoder would announce the support but
would not be able to decode a single frame ever.