transcode: forward the sout_stream_t object
The object is needed to forward the elementary streams and data to the
sout stream components following the transcode.

Note that having the stream object here means that we must ensure we
don't call it after the transcode is supposed to be closed, so we must
ensure every asynchronous components are drained and destroyed before
leaving the close functions.

The constraint was already there though, given p_obj actually pointed to
the p_stream object too.