transcode: video: refactor initialization to format update
Some decoders are pushing format update synchronously during their
activation but others have an asynchronous behaviour and may push it
before their activation function has finished, or even after the first
decode call when the decoder is asynchronously configured and ready.

Thus, previously, the decoder format from the decoder was not
necessarily known when opening the filters and the encoder. This
patchset shifts the initialization of previously mentioned objects to
the decoder update callback.

In addition, encoders will now need to get the final video context to
configure and it cannot open correctly without it, so the encoder test
typically makes no sense anymore for video. In this patch, since we
create the encoder only when the final format is finally supplied by the
decoder, we can directly create the encoder that will be used for
transcoding.

In addition, since format update would still happen at activation for
the synchronous decoders, the format and encoders will already be
available, effectively replacing the encoder test afterwards.