transcode: fix asynchronous decoder encoding
Some decoders have the decoder_QueueVideo called asynchronously wrt. the
decode() call, which is allowed and expected by the API. The MediaCodec
decoder also has a limited number of input buffers and output buffers
when mediacodec-dr is enabled.

This creates a strong backpressure linking the output buffers to the
input ones, meaning that no input buffer can be decoded unless the
already submitted output buffers are processed and released.

By running the decode() call followed by the processing of output
buffers at the same location, we're linking the output buffers
processing to the execution of the decode() call: no processing of
outputs can happen until decode() hasn't returned.

Except that decode() can be called to saturate the input buffers while
the output was already saturated, leading it to a non deterministic
deadlock.

By processing the buffer synchronously with the decoder_QueueVideo()
call, we ensure that the processing of the picture happens regardless of
the decode() blocking state. It will typically call the filter and
encode callbacks from the thread of the asynchronous module or from the
decode callback itself for synchronous decoders, leading to either the
same behaviour or difference wrt. depth-first vs breadth-first
processing at the decoder level (do we decode all the pictures we can or
do we process every new picture before decoding the others).

The refactor is made so that the actual processing function is isolated
and can be reused in a latter refactor involving a new thread if needed
to bring those difference in execution.

The previous transcode function is still handling the decoder and
encoder drain(), waiting for the decoder to queue all the video frames
and then waiting for the encoder to output every packets.

It fixes a systematic non-deterministic deadlock happening when having
the MediaCodec decoder in direct rendering mode within the transcode
pipeline, when implementing #25153.