qt: medialib: handle events from the UI thread
The medialib events notify changes that must be reflected in the UI, but
they can be executed from any thread, so it is incorrect to update the
UI from there (for example calling dataChanged() using an index not
necessary valid within the UI thread).

The callbacks provide an event structure pointing to data which are only
valid during the callback, so create a struct to copy the values and
dispatch the events to the UI thread.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>